fix: id를 Long으로 수정

+++ app/src/main/java/woowacourse/shopping/model/data/repository/ProductRepositoryImpl.kt
@@ -31,7 +31,7 @@ class ProductRepositoryImpl(
     }
 
     override fun getProductById(
-        index: Int,
+        index: Long,
         onSuccess: (Product) -> Unit,
         onFailure: (Exception) -> Unit
     ) {
