+++ domain/src/main/java/com/shopping/repository/ProductRepository.kt
@@ -10,7 +10,7 @@ interface ProductRepository {
     )
 
     fun getProductById(
-        index: Int,
+        index: Long,
         onSuccess: (Product) -> Unit,
         onFailure: (Exception) -> Unit
     )
